改修了远程和本地对“队伍等级”的保存机制。

+++ Assets/StagesManagement/FightInfo.cs
@@ -68,7 +68,7 @@ public class FightInfo : ScriptableObject
     {
         fightMembers = FightMembers.LoadEnemies_Json(Script);
         if (fightMembers != null)
-            fightMembers.SetEnemyLevel((float)stageLevel);
+            fightMembers.SetEnemyLevel(stageLevel);
     }
 
     public void LoadMyTeam()
